+++ src/test/java/com/forever/dadamda/controller/UserControllerTest.java
@@ -1,11 +1,6 @@
 package com.forever.dadamda.controller;
 
-import com.forever.dadamda.entity.user.Provider;
-import com.forever.dadamda.entity.user.Role;
-import com.forever.dadamda.entity.user.User;
 import com.forever.dadamda.mock.WithCustomMockUser;
-import com.forever.dadamda.repository.UserRepository;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
@@ -15,6 +10,8 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.TestPropertySource;
+import org.springframework.test.context.jdbc.Sql;
+import org.springframework.test.context.jdbc.Sql.ExecutionPhase;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
 
@@ -22,26 +19,12 @@ import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
 @ActiveProfiles("test")
 @AutoConfigureMockMvc
 @TestPropertySource(locations = "classpath:application-test.yml")
+@Sql(scripts = "classpath:truncate.sql", executionPhase = ExecutionPhase.AFTER_TEST_METHOD)
+@Sql(scripts = "classpath:setup.sql", executionPhase = ExecutionPhase.BEFORE_TEST_METHOD)
 public class UserControllerTest {
 
     @Autowired
     private MockMvc mockMvc;
-
-    @Autowired
-    private UserRepository userRepository;
-
-    @BeforeEach
-    void setUp() {
-        User user = User.builder()
-                .name("koko")
-                .provider(Provider.GOOGLE)
-                .role(Role.USER)
-                .email("1234@naver.com")
-                .profileUrl("https://www.naver.com")
-                .build();
-
-        userRepository.save(user);
-    }
 
     @Test
     @WithCustomMockUser
