+++ src/test/resources/truncate.sql
@@ -1,0 +1,4 @@
+SET REFERENTIAL_INTEGRITY FALSE;
+TRUNCATE TABLE scrap;
+TRUNCATE TABLE users;
+SET REFERENTIAL_INTEGRITY TRUE;
